1.fixed some known bugs(Now the device can send data to the receiver, but the receiver cannot receive data. It may be caused by the receiver's code. I had given up it.)

+++ src/UltrasonicSensor.cpp
@@ -14,6 +14,6 @@ float UltrasonicSensor::getDistance()
     delayMicroseconds(10);
     digitalWrite(trigPin, LOW);
     long duration = pulseIn(echoPin, HIGH);
-    long distance = duration * 0.034 / 2;
+    float distance = duration * 0.034 / 2;
     return distance;
 }
